chore: implement RasterLayer in TypeScript
