+++ docs/RasterLayer.md
@@ -1,4 +1,4 @@
-<!-- This file was autogenerated from RasterLayer.js do not modify -->
+<!-- This file was autogenerated from RasterLayer.tsx do not modify -->
 # <MapboxGL.RasterLayer />
 
 
@@ -11,10 +11,10 @@
 | aboveLayerID | `string` | `none` | `false` | Inserts a layer above aboveLayerID. |
 | belowLayerID | `string` | `none` | `false` | Inserts a layer below belowLayerID |
 | layerIndex | `number` | `none` | `false` | Inserts a layer at a specified index |
-| filter | `array` | `none` | `false` | Filter only the features in the source layer that satisfy a condition that you define |
+| filter | `FilterExpression` | `none` | `false` | Filter only the features in the source layer that satisfy a condition that you define |
 | minZoomLevel | `number` | `none` | `false` | The minimum zoom level at which the layer gets parsed and appears. |
 | maxZoomLevel | `number` | `none` | `false` | The maximum zoom level at which the layer gets parsed and appears. |
-| style | `union` | `none` | `false` | Customizable style attributes |
+| style | `RasterLayerStyleProps` | `none` | `true` | Customizable style attributes |
 
 
 ## styles
